requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.13.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
bokeh==3.7.3
boto3==1.34.34
botocore==1.34.162
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.3
dask==2025.7.0
distributed==2025.7.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.109.2
filelock==3.13.1
frozenlist==1.7.0
fsspec==2024.6.1
greenlet==3.2.3
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.4
jiter==0.10.0
jmespath==1.0.1
locket==1.0.0
lz4==4.4.4
Mako==1.3.10
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
narwhals==2.0.1
networkx==3.3
numpy==1.26.4
openai==1.97.1
packaging==25.0
pandas==2.2.0
partd==1.4.2
passlib==1.7.4
pillow==11.3.0
propcache==0.3.2
psutil==7.0.0
psycopg2-binary==2.9.9
pyarrow==21.0.0
pyasn1==0.6.1
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
rsa==4.9.1
s3transfer==0.10.4
safetensors==0.5.3
schedule==1.2.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.25
starlette==0.36.3
sympy==1.13.3
tblib==3.1.0
tokenizers==0.15.2
toolz==1.0.0
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.37.2
twilio==8.12.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.27.1
xyzservices==2025.4.0
yarl==1.20.1
zict==3.0.0
zipp==3.23.0
nltk==3.9.1
opensearch-py==3.0.0